@@ -7,57 +7,76 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "serve": "node serve-script",
+    "compile:server": "webpack --config webpack.server.config.js --progress --colors",
+    "serve:ssr": "node dist/server",
+    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
+    "build:client-and-server-bundles": "ng build --prod && ng run fastudio:server:production"
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular/animations": "^7.1.2",
-    "@angular/cdk": "^7.1.1",
-    "@angular/cli": "~7.1.2",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@angular-devkit/build-angular": "~0.801.3",
+    "@angular/animations": "^8.1.3",
+    "@angular/cdk": "^8.1.2",
+    "@angular/cli": "~8.1.3",
+    "@angular/common": "~8.1.3",
+    "@angular/compiler": "~8.1.3",
+    "@angular/compiler-cli": "~8.1.3",
+    "@angular/core": "~8.1.3",
+    "@angular/forms": "~8.1.3",
+    "@angular/material": "^8.1.2",
+    "@angular/platform-browser": "~8.1.3",
+    "@angular/platform-browser-dynamic": "~8.1.3",
+    "@angular/platform-server": "~8.1.3",
+    "@angular/router": "~8.1.3",
+    "@nestjs/common": "^6.0.0",
+    "@nestjs/core": "^6.0.0",
+    "@nestjs/ng-universal": "^2.0.0",
+    "@nestjs/platform-express": "^6.0.0",
+    "@nestjs/testing": "^6.5.3",
+    "@nguniversal/express-engine": "^8.1.1",
+    "@nguniversal/module-map-ngfactory-loader": "^8.0.0",
     "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
     "jwt-decode": "^2.2.0",
-    "lodash": "^4.17.11",
+    "lodash": "^4.17.15",
     "material-icons": "^0.2.3",
     "moment": "^2.23.0",
     "ng2-file-upload": "^1.3.0",
     "ng2-search-filter": "^0.4.7",
     "ng6-socket-io": "^1.0.3",
-    "ngx-cookie-service": "^2.1.0",
+    "ngx-cookie-service": "^2.2.0",
     "ngx-pagination": "^3.2.1",
     "ngx-responsive": "^7.0.1",
-    "rxjs": "~6.3.3",
+    "reflect-metadata": "^0.1.13",
+    "rxjs": "~6.5.2",
     "socket.io-client": "^2.2.0",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/language-service": "~7.1.0",
-    "@types/node": "~8.9.4",
+    "@angular/language-service": "~8.1.3",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
+    "karma": "^4.2.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
+    "nodemon": "^1.18.11",
     "protractor": "~5.4.0",
+    "rimraf": "^2.6.3",
+    "ts-loader": "^5.2.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.5",
+    "wait-on": "^3.2.0",
+    "webpack-cli": "^3.1.0"
   },
   "engines": {
     "node": "8.11.2",
