@@ -1,12 +1,15 @@
 import { Component, OnInit, ViewChild, Inject } from '@angular/core';
 import { Router } from '@angular/router';
-import { MatPaginator, MatSnackBar, MatDialog, MatDialogRef, MAT_DIALOG_DATA, MatTableDataSource } from '@angular/material';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatTableDataSource } from '@angular/material/table';
 import { HttpErrorResponse } from '@angular/common/http';
 import * as io from 'socket.io-client';
 import _ from 'lodash';
-import { AuthService } from 'src/app/shared/services/auth.service';
-import { UsersService } from 'src/app/shared/services/users.service';
-import { environment } from 'src/environments/environment';
+import { AuthService } from '../shared/services/auth.service';
+import { UsersService } from '../shared/services/users.service';
+import { environment } from '../../environments/environment';
 import { User } from '../shared/models/user';
 
 @Component({
@@ -30,7 +33,7 @@ export class DesignersComponent implements OnInit {
               public dialog: MatDialog
             ) { this.socket = io(environment.BASE_URL); }
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
+  @ViewChild(MatPaginator, { static: false }) paginator: MatPaginator;
   // @ViewChild(MatSort) sort: MatSort;
 
   ngOnInit() {
