@@ -6,7 +6,7 @@ import {
   DeleteEnquiryComponent
 } from './components/enquiries-listing/enquiries-listing.component';
 import { EnquiriesDetailsComponent } from './components/enquiries-details/enquiries-details.component';
-import { SharedModule } from 'src/app/shared/shared.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [
