@@ -1,14 +1,15 @@
 import { Component, OnInit, Inject } from '@angular/core';
 import { HttpErrorResponse } from '@angular/common/http';
-import { MatDialog, MatDialogRef, MAT_DIALOG_DATA, MatSnackBar } from '@angular/material';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router, ActivatedRoute } from '@angular/router';
 import { DesignFormComponent } from '../design-form/design-form.component';
 import * as io from 'socket.io-client';
 import * as moment from 'moment';
-import { AuthService } from 'src/app/shared/services/auth.service';
-import { UsersService } from 'src/app/shared/services/users.service';
-import { DesignService } from 'src/app/shared/services/design.service';
-import { environment } from 'src/environments/environment';
+import { AuthService } from '../../../../shared/services/auth.service';
+import { UsersService } from '../../../../shared/services/users.service';
+import { DesignService } from '../../../../shared/services/design.service';
+import { environment } from '../../../../../environments/environment';
 
 @Component({
   selector: 'app-designs',
